teuthology-describe-tests: rename from teuthology-tree
